+++ TReadoutBoardDAQ.h
@@ -125,8 +125,8 @@ class TReadoutBoardDAQ : public TUSBBoard, public TReadoutBoard {
   int  WriteRegister     (uint16_t address, uint32_t value);
 
   // DAQ board has only one control interface -> both methods are identical
-  int  SendOpCode        (uint8_t  OpCode) ;
-  int  SendOpCode        (uint8_t  OpCode, uint8_t chipId) {SendOpCode (OpCode);};
+  int  SendOpCode        (uint16_t  OpCode) ;
+  int  SendOpCode        (uint16_t  OpCode, uint8_t chipId) {SendOpCode (OpCode);};
 
   int  SetTriggerConfig  (bool enablePulse, bool enableTrigger, int triggerDelay, int pulseDelay);
   void SetTriggerSource  (TTriggerSource triggerSource);
